+++ .vscode/settings.json
@@ -2,6 +2,6 @@
     "rust-analyzer.cargo.target": "aarch64-unknown-none",
     "rust-analyzer.checkOnSave.target": "aarch64-unknown-none",
     "rust-analyzer.checkOnSave.allTargets": false,
-    "rust-analyzer.procMacro.enable": true,
+    "rust-analyzer.procMacro.enable": false,
     "editor.formatOnSave": true,
 }
