+++ libs/kernel_module/src/log.rs
@@ -1,0 +1,14 @@
+use crate::KernelService;
+use core::fmt;
+use log::Logger;
+
+impl Logger for &dyn KernelService {
+    fn log(&self, message: &str) -> Result<(), fmt::Error> {
+        KernelService::log(*self, message);
+        Ok(())
+    }
+}
+
+pub fn init() {
+    log::init(&*super::SERVICE);
+}
